pyproject.toml
[project]
name = "nonebot-plugin-simple-gpt"
version = "0.1.0"
description = "NoneBot2插件，临时的 AI 插件"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "nonebot2>=2.0.0",
    "httpx>=0.24.0",
    "openai",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}

[project.urls]
Homepage = "https://github.com/colasama/nonebot-plugin-simple-gpt"
Repository = "https://github.com/colasama/nonebot-plugin-simple-gpt"

